+++ src/py42/sdk/queries/alerts/filters/alert_filter.py
@@ -45,6 +45,44 @@ class RuleName(AlertQueryFilterStringField):
     _term = u"name"
 
 
+class RuleId(QueryFilterStringField):
+    """Class that filters alerts based on rule identifier."""
+
+    _term = u"ruleId"
+
+
+class RuleSource(QueryFilterStringField):
+    """Class that filters alerts based on rule source.
+
+    Available options are:
+        - :attr:`RuleSource.ALERTING`
+        - :attr:`RuleSource.DEPARTING_EMPLOYEE`
+        - :attr:`RuleSource.HIGH_RISK_EMPLOYEE`
+    """
+
+    _term = u"ruleSource"
+
+    ALERTING = u"Alerting"
+    DEPARTING_EMPLOYEE = u"Departing Employee"
+    HIGH_RISK_EMPLOYEE = u"High Risk Employee"
+
+
+class RuleType(QueryFilterStringField):
+    """Class that filters alerts based on rule type.
+
+    Available options are:
+        - :attr:`RuleType.ENDPOINT_EXFILTRATION`
+        - :attr:`RuleType.CLOUD_SHARE_PERMISSIONS`
+        - :attr:`RuleType.FILE_TYPE_MISMATCH`
+    """
+
+    _term = u"type"
+
+    ENDPOINT_EXFILTRATION = u"FedEndpointExfiltration"
+    CLOUD_SHARE_PERMISSIONS = u"FedCloudSharePermissions"
+    FILE_TYPE_MISMATCH = u"FedFileTypeMismatch"
+
+
 class Description(AlertQueryFilterStringField):
     """Class that filters alerts based on rule description text."""